Smammal Data HELP!
I started this project to play around with smammal data. At this point I have read in the data and I am now trying to pull out columns of the data that I am interested in working with, but I keep getting an error that the data doesnt exist, which it clearly does. Can anyone help me with this?